Update Spells, Update Campaign, Add Firebase, Update Auth (Big update)

+++ app/Http/Controllers/CampaignController.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Campaign;
+use App\CampaignCharacter;
+use App\Character;
 use App\Http\Resources\Campaign as CampaignResource;
-
-use Illuminate\Support\Facades\DB;
+use App\Jobs\SyncFirebase;
+use Illuminate\Http\Request;
+use Illuminate\Support\Str;
 use Tymon\JWTAuth\Facades\JWTAuth;
 
 class CampaignController extends Controller
@@ -16,10 +18,12 @@ class CampaignController extends Controller
     {
         $user = JWTAuth::parseToken()->authenticate();
 
-        return DB::table('campaign_characters')->join('campaigns', 'campaigns.id', '=', 'campaign_characters.campaign_id')
-            ->join('characters', 'characters.id', '=', 'campaign_characters.character_id')
-            ->select('campaigns.id', 'campaigns.name', 'campaigns.code')
-            ->where("characters.user_id", "=", $user->id)->get();
+        $campaigns = CampaignCharacter::select('campaigns.id', 'campaigns.name', 'campaigns.code', 'campaign_characters.owner')
+            ->join('campaigns', 'campaigns.id', '=', 'campaign_characters.campaign_id')
+            ->where("campaign_characters.user_id", "=", $user->id)
+            ->get();
+
+        return new CampaignResource($campaigns);
     }
 
     public function show($id)
@@ -28,26 +32,102 @@ class CampaignController extends Controller
     }
 
     // POST
-    public function store(Request $request)
+    public function addCampaign(Request $request)
     {
+        $user = JWTAuth::parseToken()->authenticate();
+
         $request->validate([
-            'name' => 'required|max:255',
-            'code' => 'required|max:255',
+            'name' => 'required|string|max:255',
         ]);
 
-        $campaign = Campaign::create($request->all());
+        if (!$user) return response()->json(['message' => 'Could not authenticate!'], 401);
 
-        return (new CampaignResource($campaign))
-            ->response()
-            ->setStatusCode(201);
+
+            $code = Str::random(8);
+
+            $campaign = Campaign::create([
+                'name' => $request->get('name'),
+                'code' => $code,
+            ]);
+
+        $this->dispatch(new SyncFirebase($campaign->id, 'campaigns'));
+
+
+        $campaignCharacter = CampaignCharacter::create([
+                'user_id' => $user->id,
+                'campaign_id' => $campaign->id,
+                'owner' => 1
+            ]);
+
+        return response()->json(array(
+                'id' => $campaign->id,
+                'name' => $campaign->name,
+                'code' => $campaign->code,
+                'owner' => $campaignCharacter->owner
+            ), 201);
+
+    }
+
+    public function joinCampaign(Request $request)
+    {
+        $user = JWTAuth::parseToken()->authenticate();
+        $data = $request->all();
+
+        $request->validate([
+            'code' => 'required|string|max:255',
+            'character_ids.*' => 'required|numeric|max:255',
+        ]);
+
+        $campaign = Campaign::where(
+            'code', '=', $data['code']
+        )->first();
+
+        if ($campaign) {
+            $existingCampaign = CampaignCharacter::where('campaign_id', '=', $campaign->id)->where('user_id', '=', $user->id)->first();
+        } else {
+            return response()->json(['message' => 'Could not find campaign with given code!'], 401);
+        }
+
+        if ($existingCampaign) return response()->json(['message' => 'Already joined this campaign!'], 401);
+
+        foreach ($data['character_ids'] as $character_id) {
+            if (!Character::whereId($character_id)->where('user_id', '=', $user->id)->first()) {
+                return response()->json(['message' => 'Invalid Characters provided!'], 401);
+            }
+        }
+
+        $character_ids = implode(',', $data['character_ids']);
+
+
+        $campaignCharacter = CampaignCharacter::create([
+            'user_id' => $user->id,
+            'campaign_id' => $campaign->id,
+            'character_id' => $character_ids,
+            'owner' => 0
+        ]);
+
+        return response()->json(array(
+            'id' => $campaign->id,
+            'name' => $campaign->name,
+            'code' => $campaign->code,
+            'owner' => $campaignCharacter->owner
+        ), 201);
     }
 
     // DELETE
-    public function delete($id)
+    public function deleteCampaign(Request $request)
     {
-        $campaign = Campaign::findOrFail($id);
-        $campaign->delete();
+        $campaignId = $request->all()['campaign_id'];
+        $user = JWTAuth::parseToken()->authenticate();
 
-        return response()->json(null, 204);
+        if (!$user) return response()->json(['message' => 'Could not authenticate!'], 401);
+        $campaign = CampaignCharacter::where('campaign_id', '=', $campaignId)->where('user_id', '=', $user->id)->first();
+        if (!$campaign) return response()->json(['message' => 'Invalid Character!'], 401);
+        elseif ($campaign->owner === 1) {
+            Campaign::whereId($campaignId)->delete();
+        } else {
+            $campaign->delete();
+        }
+        return response()->json(204);
     }
 }
